Use a PEP-487 style metaclass
PEP-487 proposes a generic metaclass to be included in the standard
library which will ease the customization of class creation, as it
is done in traitlets.

This is a showcase on how this will work. It adds the future stdlib
module and makes use of it. Once the stdlib will contain this
module, it will automatically use that instead.